refactor: [US-051] - Extract dna/csv_parser.py; move STORYGRAPH_TAG_TO_GENRE to dna_constants
